Fix handling of binary request bodies
Requests bodies were UTF-8 decoded (with a call to `toString`) when received and UTF-8 encoded when passed on the the server. As a consequence, non-UTF-8 bodies were damaged. The handling of response bodies does not have this problem.

The UTF-8 decoding with `toString` has been removed, which also prevents the subsequent encoding. Although, looking at the code handling response bodies, using the `pipe` method could be even better.